reduced list to one file
