+++ src/components/friends/FriendList.js
@@ -1,17 +1,18 @@
 import React, { useContext, useEffect, useState } from "react";
 import { FriendContext } from "./FriendProvider";
 import { UserContext } from "../users/UserProvider";
-import { useHistory, useParams } from "react-router-dom";
 import { FriendCard } from "./FriendCard";
 import "./Friend.css";
 
 export const FriendList = () => {
-  const { friends, getFriends, deleteFriend } = useContext(FriendContext);
+  const { friends, getFriends, deleteFriend, addFriend } = useContext(
+    FriendContext
+  );
   const { users, getUsers } = useContext(UserContext);
   const [filteredUsers, setUsers] = useState([]);
+  const [filteredNotFriendUsers, setNotFriendUsers] = useState([]);
 
-  const history = useHistory();
-
+  //delete friend
   const removeFriendship = (id) => {
     friends.map((friend) => {
       if (friend.userId === id) {
@@ -21,6 +22,13 @@ export const FriendList = () => {
         deleteFriend(friend.id);
       }
     });
+  };
+
+  //add friends
+  const addFriendship = (id) => {
+    const currentUser = parseInt(localStorage.user);
+    addFriend({ userId: id, followingId: currentUser });
+    addFriend({ userId: currentUser, followingId: id });
   };
 
   useEffect(() => {
@@ -39,7 +47,14 @@ export const FriendList = () => {
         followingId.includes(user.id) && user.id !== parseInt(localStorage.user)
     );
 
+    const nonFriendInformation = users.filter(
+      (user) =>
+        followingId.includes(user.id) === false &&
+        user.id !== parseInt(localStorage.user)
+    );
+
     setUsers(friendInformation);
+    setNotFriendUsers(nonFriendInformation);
   }, [friends, users]);
 
   return (
@@ -62,6 +77,29 @@ export const FriendList = () => {
               </button>
             }
           />
+        ))}
+      </div>
+      <h2>Add Friends</h2>
+      <div className="friends">
+        {filteredNotFriendUsers.map((user) => (
+          <>
+            <FriendCard
+              key={user.id}
+              friend={user}
+              isFriend={
+                <button
+                  type="submit"
+                  onClick={(e) => {
+                    e.preventDefault();
+                    addFriendship(user.id);
+                  }}
+                >
+                  {" "}
+                  Add{" "}
+                </button>
+              }
+            />
+          </>
         ))}
       </div>
     </>
